# Do less work and create less garbage when instantiating components via reflection