This partially addresses #769 by removing allocations  and unnecessary work from the hot path of reflection-based component activation.

```ini
Host Process Environment Information:
BenchmarkDotNet.Core=v0.9.9.0
OS=Windows
Processor=?, ProcessorCount=8
Frequency=2533308 ticks, Resolution=394.7408 ns, Timer=TSC
CLR=CORE, Arch=64-bit ? [RyuJIT]
GC=Concurrent Workstation
dotnet cli version: 1.0.0-preview2-003121
Type=DeepGraphResolveBenchmark  Mode=Throughput
```

### Before

|  Method |     Median |    StdDev |
|-------- |----------- |---------- |
| Resolve | 38.4004 us | 0.8659 us |


### After

|  Method |     Median |    StdDev |
|-------- |----------- |---------- |
| Resolve | 17.9659 us | 0.1156 us |