@alexmg 2017-01-03
Great stuff @nblumhardt. It seems the `ReflectionActivator` is always right in the middle of performance issues. 😃