@nblumhardt 2017-01-03
Thanks Alex. Even though the design makes some allocations unavoidable (without public API changes, anyway) I think there are still some opportunities to improve elsewhere. I'm looking forward to digging in some more when I get some time :-)