@adamsitnik 2017-01-11
@nblumhardt BenchmarkDotNet v0.10.1 has new MemoryDiagnoser, I recommend you to use it if you want to measure the number of allocated bytes as part of your benchmarking process.