@nblumhardt 2017-01-11
@adamsitnik thanks. I checked it out when I put this together, but it'd mean moving the project forward to the new `dotnet` SDK version so I couldn't get it in with this change in the time I had.